move processor tests to their own file, and add more
